feat: add TabsPanelPreview component with various tab configurations and edge cases
