+++ test/components/previews/pathogen/tabs_panel_preview/all_selected.html.erb
