+++ test/components/previews/pathogen/tabs_panel_preview/default.html.erb
