+++ test/components/previews/pathogen/tabs_panel_preview/focus_management.html.erb
