+++ test/components/previews/pathogen/tabs_panel_preview/many_tabs.html.erb
