+++ test/components/previews/pathogen/tabs_panel_preview/single_tab.html.erb
