+++ test/components/previews/pathogen/tabs_panel_preview/with_count.html.erb
