+++ test/components/previews/pathogen/tabs_panel_preview/with_icons.html.erb
